+++ src/Frustum.cpp
@@ -32,8 +32,6 @@
 #include <wendy/AABB.h>
 #include <wendy/Frustum.h>
 
-#include <cmath>
-
 ///////////////////////////////////////////////////////////////////////
 
 namespace wendy
@@ -130,8 +128,8 @@ void Frustum::set(float FOV, float aspectRatio, float farZ)
   planes.clear();
   planes.reserve(5);
 
-  const float radians = FOV * (float) M_PI / 180.f;
-  const float distance = 0.5f / tanf(radians / 2.f);
+  const float radians = FOV * float(M_PI) / 180.f;
+  const float distance = 0.5f / tan(radians / 2.f);
 
   vec3 points[5];
   points[0] = vec3(0.f);
